Fix for using mixin on `MobileScrollFixLayer`
Fixes #1.

The problem occurred when trying to use `Material.mixin` on `Layer`
on mobile because `Layer` is replaced with `MobileScrollFixLayer` ...
which isn't a property on Framer.Defaults.

This is brittle and I should probably just reconsider the mixin API.
Maybe it is worth making calls to mixin take no argument and just
extend `Layer` by default, although I like the clarity that comes
from explicitly passing in the object you want to extend.